+++ inventory-service/src/main/java/com/urbanthreads/inventoryservice/repo/ItemRepository.java
@@ -4,6 +4,8 @@ import com.urbanthreads.inventoryservice.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 }
